Improved Error management (#32)

+++ install.sh
@@ -1,27 +1,43 @@
 #!/usr/bin/env bash
 
 install() {
+	# dependencies
 	for gem in "${gems[@]}"; do
 		is_gem_installed "$gem" || sudo gem install "$gem"
 	done
 
 	[ -d "$CONF_DIR" ] || sudo mkdir "$CONF_DIR"
+	[ -d "$SCHEMAS_DIR" ] || sudo mkdir "$SCHEMAS_DIR"
 	[ -d "$INSTALL_DIR" ] || mkdir "$INSTALL_DIR"
 
 	src_conf="$(realpath "share/etc/$CONF_FILE")"
 	src_exec="$(realpath src/server/cli/provision-engine-server)"
 
 	if [[ $setup_mode == "copy" ]]; then
+		# config
 		[ -f "$CONF_PATH" ] || sudo cp "$src_conf" "$CONF_PATH"
-		[ -L "$EXEC_PATH" ] || sudo cp "$src_exec" "$EXEC_PATH"
+		for file in $SCHEMAS; do
+			[ -f "$CONF_DIR/${file}" ] || sudo cp "$(realpath "share/schemas/${file}")" "$SCHEMAS_DIR"
+		done
 
+		# executable
+		[ -f "$EXEC_PATH" ] || sudo cp "$src_exec" "$EXEC_PATH"
+
+		# libraries
 		for file in $modules; do
-			[ -L "$INSTALL_DIR/${file}" ] || cp "$(realpath "src/server/${file}")" "${INSTALL_DIR}"
+			[ -f "$INSTALL_DIR/${file}" ] || cp "$(realpath "src/server/${file}")" "${INSTALL_DIR}"
 		done
 	elif [[ $setup_mode == "symlink" ]]; then
+		# config
 		[ -L "$CONF_PATH" ] || [ -f "$CONF_PATH" ] || sudo ln -s "$src_conf" "$CONF_PATH"
+		for file in $SCHEMAS; do
+			[ -L "$CONF_DIR/${file}" ] || sudo ln -s "$(realpath "share/schemas/${file}")" "$SCHEMAS_DIR"
+		done
+
+		# executable
 		[ -L "$EXEC_PATH" ] || sudo ln -s "$src_exec" "$EXEC_PATH"
 
+		# libraries
 		for file in $modules; do
 			[ -L "$INSTALL_DIR/${file}" ] || ln -s "$(realpath "src/server/${file}")" "${INSTALL_DIR}"
 		done
@@ -36,6 +52,7 @@ postinstall() {
 clean() {
 	if [[ $setup_mode == "purge" ]]; then
 		[ -d $CONF_DIR ] && sudo rm -r $CONF_DIR
+		[ -d "$SCHEMAS_DIR" ] && sudo rm -rf "$SCHEMAS_DIR"
 
 		for gem in "${gems[@]}"; do
 			remove_orphan_gem "$gem"
@@ -45,8 +62,7 @@ clean() {
 		echo "The ruby gems: $(printf "%s " "${gems[@]}") are required by the provision engine"
 	fi
 
-	[ -f "$EXEC_PATH" ] && sudo rm $EXEC_PATH
-	[ -L "$EXEC_PATH" ] && sudo rm $EXEC_PATH
+	[ -f "$EXEC_PATH" ] || [ -L "$EXEC_PATH" ] && sudo rm $EXEC_PATH
 	[ -d "$INSTALL_DIR" ] && rm -rf "$INSTALL_DIR"
 }
 
@@ -61,11 +77,15 @@ remove_orphan_gem() {
 CONF_DIR="/etc/provision-engine"
 CONF_FILE="engine.conf"
 CONF_PATH="${CONF_DIR}/${CONF_FILE}"
+SCHEMAS_DIR="/etc/provision-engine/schemas"
+SCHEMAS="serverless_runtime.json error.json"
+
 EXEC_FILE="provision-engine-server"
 EXEC_PATH="/usr/local/bin/${EXEC_FILE}"
-INSTALL_DIR="/opt/provision-engine" # install location is hardcoded
 
-modules="client.rb configuration.rb log.rb server.rb runtime.rb"
+INSTALL_DIR="/opt/provision-engine"
+modules="client.rb configuration.rb log.rb server.rb runtime.rb error.rb"
+
 gems=("opennebula" "sinatra" "logger" "json-schema") # check requires on server.rb
 
 action="${1:-"install"}"
